git-svn-id: svn://svn.appwork.org/utils@676 21714237-3853-44ef-a1f0-ef8f03a7d1fe

+++ src/org/appwork/utils/net/throttledconnection/ThrottledConnection.java
@@ -19,7 +19,7 @@ public interface ThrottledConnection {
      * 
      * @return
      */
-    public long getCustomLimit();
+    public int getCustomLimit();
 
     /**
      * sets custom speed limit -1 : no limit 0 : use managed limit >0: use
